remove unused modules
