+++ language/src/Circuit/Language.hs
@@ -1,10 +1,10 @@
 module Circuit.Language
   ( module Circuit.Language.Compile,
     module Circuit.Language.DSL,
-    module Circuit.Language.TExpr,
+    module Circuit.Language.Expr,
   )
 where
 
 import Circuit.Language.Compile
 import Circuit.Language.DSL
-import Circuit.Language.TExpr
+import Circuit.Language.Expr
